+++ ischeduler/ischeduler.go
@@ -1,7 +1,7 @@
 package ischeduler
 
 type IScheduler interface {
-	AddRun(run int)
-	AddRunner(runner int)
+	AddRun(run int64)
+	AddRunner(runner int64)
 	Start()
 }
