+++ message/messages.go
@@ -13,24 +13,24 @@ const (
 type SubscribeRequest struct {
 	Token       string
 	Name        string
-	Concurrency int
+	Concurrency int64
 }
 
 type SubscribeAnswer struct {
-	ClientUID int
+	ClientUID int64
 }
 
 type ClientRequest struct {
-	Kind    int
-	JobID   int
+	Kind    int64
+	JobID   int64
 	Logs    []string
 	Message string
 	Status  string
 }
 
 type ServerRequest struct {
-	Kind  int
-	JobID int
+	Kind  int64
+	JobID int64
 	Run   *RunRequest
 }
 
@@ -38,6 +38,6 @@ type RunRequest struct {
 	Init       string
 	Repo       string
 	Name       string
-	UpdateTime uint
-	Timeout    uint
+	UpdateTime uint64
+	Timeout    uint64
 }
